backend/requirements.txt
pandas
numpy
opencv-python
tensorflow
scikit-learn
matplotlib
seaborn
jupyterlab
fastapi
uvicorn
sqlalchemy
pydantic
passlib[bcrypt]
python-jose
python-multipart